# [DO NOT MERGE] Measures performance with normal parser