## With normal parser
```
simple query - validate syntax x 6,449 ops/sec ±0.94% (29 runs sampled)
tictactoe query - validate syntax x 82.27 ops/sec ±2.48% (24 runs sampled)
large query - validate syntax x 3.62 ops/sec ±2.24% (11 runs sampled)
```

## With webworker

```
simple query - validate syntax x 257 ops/sec ±2.07% (11 runs sampled)
tictactoe query - validate syntax x 18.14 ops/sec ±3.63% (9 runs sampled)
For the large query I could only do one run - 0.0208 ops/sec
```

So overall:

```
x25 times slower
x4.53 times slower
x174 times slower
```